make new file for napi5 static methods
